rename withStudyNotContributor
