+++ src/app/(dashboard)/etudes/[id]/(organization)/cadrage/ajouter-contributeur/page.tsx
@@ -1,10 +1,10 @@
 import withAuth from '@/components/hoc/withAuth'
 import { StudyProps } from '@/components/hoc/withStudy'
-import withStudyNotContributor from '@/components/hoc/withStudyNotContributor'
+import withStudyDetails from '@/components/hoc/withStudyDetails'
 import NewStudyContributorPage from '@/components/pages/NewStudyContributor'
 
 const NewStudyContributor = async ({ study }: StudyProps) => {
   return <NewStudyContributorPage study={study} />
 }
 
-export default withAuth(withStudyNotContributor(NewStudyContributor))
+export default withAuth(withStudyDetails(NewStudyContributor))
