+++ src/app/(dashboard)/etudes/[id]/(organization)/cadrage/ajouter/page.tsx
@@ -1,10 +1,10 @@
 import withAuth, { UserProps } from '@/components/hoc/withAuth'
 import { StudyProps } from '@/components/hoc/withStudy'
-import withStudyNotContributor from '@/components/hoc/withStudyNotContributor'
+import withStudyDetails from '@/components/hoc/withStudyDetails'
 import NewStudyRightPage from '@/components/pages/NewStudyRight'
 
 const NewStudyRight = async (props: StudyProps & UserProps) => {
   return <NewStudyRightPage study={props.study} user={props.user} />
 }
 
-export default withAuth(withStudyNotContributor(NewStudyRight))
+export default withAuth(withStudyDetails(NewStudyRight))
