+++ src/app/(dashboard)/etudes/[id]/(organization)/cadrage/page.tsx
@@ -1,6 +1,6 @@
 import withAuth, { UserProps } from '@/components/hoc/withAuth'
 import { StudyProps } from '@/components/hoc/withStudy'
-import withStudyNotContributor from '@/components/hoc/withStudyNotContributor'
+import withStudyDetails from '@/components/hoc/withStudyDetails'
 import StudyRightsPage from '@/components/pages/StudyRights'
 
 export const revalidate = 0
@@ -9,4 +9,4 @@ const StudyRights = async (props: StudyProps & UserProps) => {
   return <StudyRightsPage study={props.study} user={props.user} />
 }
 
-export default withAuth(withStudyNotContributor(StudyRights))
+export default withAuth(withStudyDetails(StudyRights))
