+++ src/app/(dashboard)/etudes/[id]/(organization)/comptabilisation/resultats/page.tsx
@@ -1,6 +1,6 @@
 import withAuth from '@/components/hoc/withAuth'
 import { StudyProps } from '@/components/hoc/withStudy'
-import withStudyNotContributor from '@/components/hoc/withStudyNotContributor'
+import withStudyDetails from '@/components/hoc/withStudyDetails'
 import ResultsPage from '@/components/pages/Results'
 import { getEmissionFactorsWithPartsInIds } from '@/db/emissionFactors'
 import { getExportRules } from '@/db/exportRule'
@@ -14,4 +14,4 @@ const ResultatsPages = async ({ study }: StudyProps) => {
   return <ResultsPage study={study} rules={rules} emissionFactorsWithParts={emissionFactorsWithParts} />
 }
 
-export default withAuth(withStudyNotContributor(ResultatsPages))
+export default withAuth(withStudyDetails(ResultatsPages))
