+++ src/app/(dashboard)/etudes/[id]/(organization)/comptabilisation/saisie-des-donnees/page.tsx
@@ -1,10 +1,10 @@
 import withAuth from '@/components/hoc/withAuth'
 import { StudyProps } from '@/components/hoc/withStudy'
-import withStudyNotContributor from '@/components/hoc/withStudyNotContributor'
+import withStudyDetails from '@/components/hoc/withStudyDetails'
 import StudyContributionPage from '@/components/pages/StudyContribution'
 
 const DataEntry = async ({ study }: StudyProps) => {
   return <StudyContributionPage study={study} />
 }
 
-export default withAuth(withStudyNotContributor(DataEntry))
+export default withAuth(withStudyDetails(DataEntry))
