+++ src/app/(dashboard)/etudes/[id]/(organization)/page.tsx
@@ -1,10 +1,10 @@
 import withAuth from '@/components/hoc/withAuth'
 import { StudyProps } from '@/components/hoc/withStudy'
-import withStudyNotContributor from '@/components/hoc/withStudyNotContributor'
+import withStudyDetails from '@/components/hoc/withStudyDetails'
 import StudyPage from '@/components/pages/Study'
 
 const StudyView = async ({ study }: StudyProps) => {
   return <StudyPage study={study} />
 }
 
-export default withAuth(withStudyNotContributor(StudyView))
+export default withAuth(withStudyDetails(StudyView))
