+++ src/app/(dashboard)/etudes/[id]/(organization)/perimetre/page.tsx
@@ -1,10 +1,10 @@
 import withAuth, { UserProps } from '@/components/hoc/withAuth'
 import { StudyProps } from '@/components/hoc/withStudy'
-import withStudyNotContributor from '@/components/hoc/withStudyNotContributor'
+import withStudyDetails from '@/components/hoc/withStudyDetails'
 import StudyPerimeterPage from '@/components/pages/StudyPerimeter'
 
 const StudyPerimeter = async (props: StudyProps & UserProps) => {
   return <StudyPerimeterPage study={props.study} user={props.user} />
 }
 
-export default withAuth(withStudyNotContributor(StudyPerimeter))
+export default withAuth(withStudyDetails(StudyPerimeter))
