+++ src/components/hoc/withStudyDetails.tsx
